Add an interface and base class for cluster manager.
